requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.24.0
mysql-connector==2.2.9
numpy==1.19.2
PyMySQL==0.10.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
Werkzeug==1.0.1